Correct namespaces and little refactor
