+++ GMCS_RestAPI/Controllers/BooksController.cs
@@ -3,9 +3,10 @@ using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using GMCS_RestAPI.Classes;
-using GMCS_RestAPI.Database;
-using GMCS_RestAPI.Models;
+using GMCS_RestApi.Domain.Classes;
+using GMCS_RestApi.Domain.Contexts;
+using GMCS_RestApi.Domain.Enums;
+using GMCS_RestApi.Domain.Models;
 using Microsoft.EntityFrameworkCore;
 
 namespace GMCS_RestAPI.Controllers
@@ -14,7 +15,7 @@ namespace GMCS_RestAPI.Controllers
 	[Route("[controller]")]
 	public class BooksController : ControllerBase
 	{
-		private ApplicationContext _context;
+		private readonly ApplicationContext _context;
 
 		public BooksController(ApplicationContext context)
 		{
@@ -115,7 +116,7 @@ namespace GMCS_RestAPI.Controllers
 				return BadRequest();
 			}
 
-			book.BookStateId = 2;
+			book.BookStateId = (int)EBookState.InStock;
 
 			book.WhoChanged = Environment.UserName;
 
@@ -142,12 +143,12 @@ namespace GMCS_RestAPI.Controllers
 				return BadRequest();
 			}
 
-			if (book.BookStateId != 2)
+			if (book.BookStateId != (int)EBookState.InStock)
 			{
 				return BadRequest("Данной книги нет в налчии");
 			}
 
-			book.BookStateId = 1;
+			book.BookStateId = (int)EBookState.Sold;
 
 			book.WhoChanged = Environment.UserName;
 
@@ -176,7 +177,7 @@ namespace GMCS_RestAPI.Controllers
 				return BadRequest(ModelState);
 			}
 
-			_context.Books.Add(book);
+			await _context.Books.AddAsync(book);
 			await _context.SaveChangesAsync();
 			return Ok(book);
 		}
