+++ GMCS_RestApi.Domain/Enums/EBookState.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace GMCS_RestApi.Domain.Enums
+{
+    public enum EBookState
+    {
+        Sold = 1,
+        InStock = 2,
+        Unknown = 3
+    }
+}
