+++ GMCS_RestApi.Domain/Models/Book.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace GMCS_RestAPI.Models
+namespace GMCS_RestApi.Domain.Models
 {
 	public class Book
 	{
