+++ GMCS_RestApi.Domain/Models/BookState.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace GMCS_RestAPI.Models
+namespace GMCS_RestApi.Domain.Models
 {
 	public class BookState
 	{
